Fix AMP issue when processing multi-assets (#16344)
Summary:
In https://github.com/dagster-io/dagster/pull/16150 I didn't realize
that the ToposortPriorityQueue also includes any sibling multi-asset
keys in the keys that we iterate through. The previous recursive logic
didn't do that, and we don't need to either here - we just need to sort
the passed in keys topologically. In fact, doing so messes things up
since we end up running on keys that never made it into the cache.

Test Plan: working on a unit test case that was failing before - running
a dry run on a specific org that was hitting a KeyError due to a
multi-asset no longer hits problems

## Summary & Motivation

## How I Tested These Changes