[ui] Some backfill page cleanup (#16310)
## Summary & Motivation

A handful of changes and cleanup items on the backfill table and
backfill page.

- Virtualize the partition and asset key tables from these pages,
reusing the component created for auto-materialization. It's just a
generic virtualized item list anyway, so we can go ahead and reuse it
here. (I moved and renamed the component to give it generic
characteristics.)
- When linking to an asset from the backfill page where a backfill is
for a range, link to that default range on the asset page itself.
- When a backfill header just shows the names of partitions that have
been backfilled (with no dialog) use a `Tag` to give the partition names
some separation. It's not perfect, but it seemed a little weird to just
stick a comma between them.
- Some alignment tweaks.

<img width="557" alt="Screenshot 2023-09-05 at 9 51 06 AM"
src="https://github.com/dagster-io/dagster/assets/2823852/e70676ab-e89b-4b36-89d2-8ad2fe337dc2">
<img width="580" alt="Screenshot 2023-09-05 at 9 50 58 AM"
src="https://github.com/dagster-io/dagster/assets/2823852/c995a16f-7345-4a9f-a22e-2031e269b661">

## How I Tested These Changes

View Backfill page. Click to view assets affected by a backfill, verify
dialog with virtualized contents. Verify same when clicking the link to
see the affected partitions.

Click into a backfill, verify same dialog and virtualization behavior
when looking at the list of affected partitions.

Click into a backfill that uses a partition range. Click to the asset,
verify that the range is applied to the querystring of the URL and loads
as the selected range on the asset page.