Make all ExtClients plain-ole-python objects (#16352)
## Summary & Motivation

Making this class hierarchy not inherit from `ConfigurableResource`. To avoid forcing users from using `ResourceParam`, with every use, I am proposing a pattern where we export the annotated type. We could support this more directly and ergonomically in the framework, but I think this the user-facing behavior we want for ext users.

## How I Tested These Changes

BK