[RFC][Overview Assets] Fetching status for 1 asset at a time. (#16245)
## Summary & Motivation

A certain Dagster user has thousands of assets inside of a single group
and currently we're fetching all of the assets of a group at once. This
PR changes the logic to instead fetch a single asset at a time per
group.

This is a bandaid of sorts for the time being. Will deploy this to
dogfood to get a better sense of how it performs for our dogfood test
orgs. I tried fetching in bigger batches of 1 but some of the queries
were still timing out. I'm consdering making this configurable via an
org setting.

## How I Tested These Changes
Shadow dagit (which is extremely slow)

<img width="1494" alt="Screenshot 2023-08-31 at 2 15 17 PM"
src="https://github.com/dagster-io/dagster/assets/2286579/3b17ebd8-70ef-4dd5-b85a-7ef9761fb09a">