Allow OpExecutionContext.add_output_metadata to be called multiple times per output (#16391)
## Summary & Motivation

`OpExecutionContext.add_output_metadata` imposes a restriction that it
can only be called once-per-output. It's unclear to me what the reason
for this restriction is, and it complicates the implementation of
metadata reporting in `dagster-ext`. This PR removes the restriction.

cc @dpeng817 since it looks like you added `add_output_metadata` way
back in #6530

## How I Tested These Changes

Modified existing unit tests.