Show asset checks if they were in the same run as a materialization (#16439)
Previously, we hid checks if they returned a result before the latest
materialization. This means we hide them with the asset promotion
pattern. This diff shows the check as long as it was evaluated in the
same run as the eventual materialization.

I suspect we'll need to iterate here. E.g. if I run the promotion step
and not the check, i'll still hide the check. We might end up with a
flag on checks to indicate that they apply to future materializations. I
think if we ship this simple change, it will let users get far enough to
give us feedback about how they want check status shown in the various
states with this pattern.

Tests:
- unit
- using the toys graph asset example