[refactor] autodoc_dagster.py -> dagster-sphinx (#15407)
## Summary & Motivation

Our custom sphinx extension `autodoc_dagster` currently lives in a
single file. This PR breaks it into a separate Python package
`dagster-sphinx` and refactors it somewhat. This was necessary due to
the growing complexity of the extension due to recent and in-progress
docs improvements.

This PR also adds the new package to the pyright master env, which
introduces `sphinx` and `docutils` dependencies and necessitate a
pyright pins update (bouns: developing a sphinx plugin is way more
pleasant with pyright/LSP help).

## How I Tested These Changes

Built docs.