Make injectors take no params and move tempdir handling to body of inject_context (#16354)
## Summary & Motivation

We want to restructure the ext clients so that they receive the context
injector arguments as `__init__` time, rather than as a parameter to
run. However right now in some cases (namely the file case) the context
injector has to be constructed when the temp directory is already
created. This moves the temp dir creation logic into the
`inject_context` method itself. This makes the `__init__` trivial. The
downside is that we create two tempdirs, one for the message reader and
one for the context injector. However that in my judgement is immateriai

## How I Tested These Changes

BK