Raise informative error when UPathIOManager cannot persist single-run backfill output (#16416)
Previously, users who launched single run backfills via the UI would
receive an obscure `assert len(paths) == 1` failure when executing with
a `UPathIOManager`. The cause of this error is that the `UPathIOManager`
cannot handle outputs that correspond to multiple partitions
(https://github.com/dagster-io/dagster/issues/14332).

For now, this PR improves the error messaging to better describe the
cause of the error.