[gql] Add query to fetch run IDs with filter (#16461)
This PR adds a `runIdsOrError` resolver that mirrors the existing
`runsOrError` resolver.

The immediate use case here is to enable terminating runs with a given
status. The client can use this resolver to quickly fetch queued run IDs
and then execute the terminate pipeline mutation on each run ID. This
allows us also to display progress as the cancellation operation occurs.