Remove asset check definition check (#16485)
This dagster._check check calls
`assets_defs_by_node_handle.get(node_handle).get_spec_for_check_handle(handle)`.

When inside a graph asset, if an op only returns a check and not an
output, then we don't have an AssetsDefinition for it in the mapping so
this check fails.

I think it's fairly safe to remove this check. We already use
get_check_names_by_asset_key_for_node_handle to get here so the check
exists. We just don't have an AssetsDefinition for it on the op node (we
do on the graph node).