[ui] Show one asset run tag or “4 assets”, hover action for lineage (#16449)
## Summary & Motivation

This PR changes the asset tags that appear in the run table and on the
run details page:

1) Rather than show 1-3 individual assets or a "6 assets" link in the
run table or run details header, we now show either a single asset or "6
assets". This means the rendering is more compact in the 2-3 asset case
since the tags are not rendered individually.

2) Instead of being directly clickable, hovering over the assets gives
you the option to jump to a lineage view OR open the list of assets. For
>1 asset, the downstream lineage option goes to the global asset graph
with the "assetA*, assetB*" query. It's actually pretty slick.

Right now in Dagit we don't have a "default behavior" for clicking tags
that have tag actions -- you have to choose an action. I could see
keeping the default "link" behavior we had before though...

Before:

![image](https://github.com/dagster-io/dagster/assets/1037212/e09c7998-d92a-496f-96b4-bcdd7d48c253)

![image](https://github.com/dagster-io/dagster/assets/1037212/ea86bac2-bf6a-4c03-84b6-3c134c01db5f)

![image](https://github.com/dagster-io/dagster/assets/1037212/44c711e0-dfef-4213-acaf-deff4f70b489)

After:

![image](https://github.com/dagster-io/dagster/assets/1037212/f81a4178-c3af-437a-bcba-1db760750244)

![image](https://github.com/dagster-io/dagster/assets/1037212/845dfd0b-b1b7-438b-8d70-b9130fba9fb1)

![image](https://github.com/dagster-io/dagster/assets/1037212/836cbb21-f485-4555-b82a-dfe6303cbeec)

Also verified with multi-component asset keys:

![image](https://github.com/dagster-io/dagster/assets/1037212/882bced3-382c-4bdb-b420-93e32be2038e)

![image](https://github.com/dagster-io/dagster/assets/1037212/da6b77b9-6ff9-40ef-aced-235fb6677439)


## How I Tested These Changes

I viewed the runs list and run details for a wide range of runs.

---------

Co-authored-by: bengotow <bgotow@elementl.com>