[dbt] Raise error on package conflicts in dagster-dbt project scaffold (#16484)
Resolves https://github.com/dagster-io/dagster/issues/16367.

Users are experiencing import errors when dagster projects are
scaffolded using names that conflict with required PyPI packages i.e.
`dagster` or `dagster_dbt`.

This PR fixes the issue by raising an error in `dagster project
scaffold` and `dagster-dbt project scaffold` CLIs when the code location
name causes a conflict with an existing package. It only checks for
existing packages if the code location name contains dagster or
dbt--this is because a lot of names are valid python packages (I.e.
hello, foo, and yay are all PyPI packages) and it seems restrictive to
block all of these words.

We raise an error when package collisions exist regardless of
hyphen/underscore:
- when dagster-dbt is provided, this could prevent the actual
dagster-dbt package from being installed
- when dagster_dbt is provided, this causes import errors in code i.e.
when from dagster_dbt import DbtCliResource is called

If users want to bypass the error, they can provide an optional
`--ignore-package-conflict` flag to bypass the package conflict check.