Test some common build scenarios
This commit adds a few example tests to demonstrate how we can use these
new fixtures to test changes to our build pipeline.

I think writing these tests is a pretty good justification for the need
for these tests - already they've unearthed inconsistencies with how we
expect our builds to work (for example, merges should only run on
python 3.10) and how they actually work (merges run dagster-airflow on
python 3.8). I've annotated this with an xfail and hope to find time to
go fix some of these oddities later.

Another thing I noticed while writing these is that we're inconsistent
about when we add a skipped step to the pipeline vs. when we don't add
the step to the pipeline at all. I think a better pattern would be to
always build the same pipeline, but to use skip annotations to tell
Buildkite which ones to run.