[ui] Link "Launched by" tags to schedules/sensors (#16502)
## Summary & Motivation

Resolves #16475.

If a RepoAddress is available for a run, use it to link the
schedule/sensor in the "Launched by" column on the Runs table.

## How I Tested These Changes

View runs launched by schedules and sensors in the Runs table. Verify
that they are properly linked.

Find some runs that are launched by schedules/sensors that no longer
exist in the workspace. Verify that these are not linked.