[auto-materialize] Change default behavior for self-dependent assets (#16492)
## Summary & Motivation

This makes three changes to how we deal with self-dependent assets with
AMP.

1. Updates to past partitions of a self-dependent asset no longer
trigger the `AutoMaterializeRule.materialize_on_parent_updated` rule for
subsequent partitions of that asset. This prevents "slow motion
backfills if a user were to update a random past partition of that asset
(kicking off a chain of computations)
2. Past partitions of a self-dependent asset are not used to determine
if a future partition of that asset is out of date or not. This means
that if a long-past partition of that asset is updated, that won't
prevent future partitions of that asset from being updated.
3. Past partitions of a self-dependent asset do not need to be updated
in order for the `AutoMaterializeRule.skip_on_not_all_parents_updated`
rule to be triggered.

This is not currently toggleable, as it's unclear if the previous
default behavior would be desirable.

## How I Tested These Changes