AssetSpec.asset_key --> AssetSpec.key (#16544)
## Summary & Motivation

A discussion in our internal slack with @johannkm and @jamiedemaria––as
well as on the ground coding––has convinced me that going to `asset_key`
in `AssetSpec` was a mistake. This PR proposes to change
`AssetSpec.asset_key` to `AssetSpec.key`.

This is consistent with `AssetsDefinition` and avoids a lot of code that
looks like `asset.asset_key` which does not feel great.



## How I Tested These Changes