[ui] Fix undefined value being passed to asset launchpad (#16565)
## Summary & Motivation

There's a TS loophole in our use of `Partial` when generating a
transient execution session for the asset launchpad.

We can inadvertently spread in an `undefined` value for the run config
yaml, which isn't caught by TS because it's a key of a `Partial` object.
TS assumes the return value from the spread is a fully formed execution
session object, with a valid `string` config yaml. This persists
downstream and surfaces in the CodeMirror editor, which expects a
`string` value and errors when provided an `undefiend`.

Repair the callsite that generates the launchpad session by avoiding the
`undefined` value.

## How I Tested These Changes

View asset job that shows this error, with `flagDisableAutoLoadDefaults`
turned on to force the launchpad not to show a default config.
Shift-click to materialize, verify that the asset launchpad loads
properly, with no JS errors.