Add shell for create_unexecutable_observable_assets_def: (#16575)
## Summary & Motivation

Add create_unexecutable_observable_assets_def. This will be used by the
internals of repository creation to create an AssetsDefinition that
represents an asset that cannot be materialized. For now it just throws
a not implemented exception. In downstream PRs we will annotate this
AssetsDefinition so that tools and calling code can know that this
cannot be materialized.

## How I Tested These Changes

BK

---------

Co-authored-by: Nicholas Schrock <schrockn@elementl.com>