Add system metadata key to determine executability (#16576)
## Summary & Motivation

For UIs or calling code to determine if an asset key is executable, we
need to add this information to `AssetsDefinition`. Rather than add a
property and to go through the associated insanity of adding said
property to >10 layers of passing around a new dictionary indexed by
assets key, I'm just going to jam it in the user-defined metadata with a
new system-level metadata key.

The metadata key is an enum that contains information about the
"varietal" of this assets definition. This allows for extension later. I
have a near-term extension in mind, as new varietal––executable
observable assets––will be reintroduced in short order to supplant
observable source assets.

We can reassess this later, but I do not want to introduce API and code
thrash for now.

This also makes `create_unexecutable_observable_assets_def` use
`multi_asset` instead of an `asset`

## How I Tested These Changes

BK

---------

Co-authored-by: Nicholas Schrock <schrockn@elementl.com>