Temporarily disable k8s-test-suite (#16606)
## Summary & Motivation


![dhh](https://github.com/dagster-io/dagster/assets/28738937/e6c5048f-c76e-405a-af1c-3ca8d6eec797)

We're just going to turn off the tests. They're no fun. The code looks
better. I know this is bound to be controversial and we'll never reach
alignment, so we're just going ahead with this.

Just kidding: AWS causing this test to be very flaky. Disable until it
is resolved.

## How I Tested These Changes

BK