[observable-source-assets] Fix observable job resolution logic (#16626)
## Summary & Motivation

Fixes: https://github.com/dagster-io/dagster/issues/16611

The prior logic made sure to appropriately match observable source
assets to asset jobs that shared an identical underlying
PartitionsDefinition. However, if your code location consists of only
partitioned assets, and your observable source asset is unpartitioned,
this means that no job can be found to materialize it.

Any unpartitioned asset can be materialized by any partitioned job, so
this PR updates the logic to allow for unpartitioned assets to be
materialized by partitioned base asset jobs.

## How I Tested These Changes

The added test case is a bit goofy, but it did reproduce the error.