[ui] Show a delayed loading spinner on Overview search inputs (#16607)
## Summary & Motivation

Show a delayed loading spinner on Overview search inputs to inform the
user that the workspace is still loading. This way, when they attempt a
search that comes up empty, there will be an obvious reason for it: the
workspace isn't ready yet.

<img width="562" alt="Screenshot 2023-09-18 at 4 59 43 PM"
src="https://github.com/dagster-io/dagster/assets/2823852/8aec765c-7ed3-4430-b623-c7a105bd2052">

I added a utility hook to `ui-components` for a delayed state update,
which will allow us to wait briefly before showing the spinner. This
prevents a quick flash of the spinner in cases where the workspace loads
fairly quickly. I'll use the utility in a couple of places in Cloud that
I've done this in recently.

## How I Tested These Changes

View Overview, verify that spinners appear after a brief delay when the
loading state is forced to be true.

Storybook example for the utility hook.