Don't set a SolidSelection tag when you materialize a set of assets
Summary:
Much like https://github.com/dagster-io/dagster/pull/13394 - this dagster/solid_selection (or dagster/op_selection) tag can be very large, but I can find no place where we actually use it anymore on the backend.

It still could be useful for non-asset jobs to see what ops you launched by showing the query - so instead, for each of the other callsites where we were setting a solidSelectionQuery, set that tags there instead.

Test Plan:
Load an asset job locally, press "materialize all" in the launchpad - all assets materialize but the giant tag is no longer there
In a non-asset job, materailize an op subset - tag is still there on the resulting job
Presets are no longer possible to set, but do the same for a partitioned-job that sets an op subset (is this still a thing?) and verify the tag is still set on the resulting job