[ui] Fix /guess paths (#16653)
## Summary & Motivation

We're currently producing `/guess` links that use `/guess/jobs` or
`/guess/pipelines`, which is incorrect. We just need
`/guess/job_or_pipeline_name`.

This becomes a problem when navigating from the tag in the run header
than indicates the pipeline/job for the run. If the workspace hasn't
loaded yet, we use the `/guess` link, which leads to a page that says
that "pipeline" is not found in the workspace.

## How I Tested These Changes

View `/guess/my_pipeline_name` URL, verify that it correctly determines
the code location and performs the redirect.