[1/n] Asset graph sidebar (#16447)
## Summary & Motivation

https://www.loom.com/share/647d76108782411f8ef1c9b06f55eae5

This adds a sidebar to the left of the asset graph that makes it easier
to navigate the asset graph without needing to scroll around. This is
useful for cases where the asset graph is very large and it's hard to
see the ancestors / descendants of a particular asset.

It would be nice if we could figure out how to combine this with the
existing asset graph sidebar that shows up on the right side somehow so
that there's only one sidebar.

I put minimal effort into the design for now since this is just a
prototype I built to demonstrate one solution to the problem of
navigating large asset graphs. Looking for input as to whether we should
pursue this direction or not.