feat(dbt): validate that `dbt-core>=1.4.0` when using `DbtCliResource` (#16702)
## Summary & Motivation
A couple of users have run into this, so rather than disclosing this in
documentation, enforce it at the framework level.

When using `DbtCliResource`, `dbt-core>=1.4.0` is required as we make
use of [dbt's standardized event log
structure](https://docs.getdbt.com/reference/events-logging) to emit
Dagster events.

Until `dbt-core<1.4` is EOL, we should emit an exception when the user
using an older version of `dbt-core`.

## How I Tested These Changes
pytest