Add require_typed_event_stream to compute contexts (#16706)
## Summary & Motivation

Per conversation with @schrockn and @rexledesma, add a
`require_typed_event_stream` switch on `OpExecutionContext`. This is off
by default and has to be explicitly turned on, which we do upstack in
`ext_protocol`. The switch has two effects:

- Skips special validation done on returned values from ops, instead
wrapping the returned result without alteration in a generator and
treating them as if they had been yielded
- Causes an error to be thrown when an expected result is missing.

The implementation here is moderately greasy and should be considered
temporary. A better solution will take time since the code paths
governing op results are complex.

## How I Tested These Changes

New unit tests.