feat(dbt): create unique target path in existing `DBT_TARGET_PATH` (#16744)
## Summary & Motivation
If the user has already specified `DBT_TARGET_PATH`, then use it as the
parent directory when creating artifact files using `DbtCliResource`.

If `DBT_TARGET_PATH` is a relative path, then continue create artifacts
as a subdirectory of the dbt project directory. Otherwise,
`DBT_TARGET_PATH` is an absolute path, and artifacts should be created
as subdirectories of that.

## How I Tested These Changes
pytest