Make @graph_asset support explicit key (#16751)
## Summary & Motivation

`graph_asset`'s parameters are out-of-sync with `asset`'s on a structural level, and we do not have the discipline or the tests to keep them in sync. This PR adds an explicit `key` which is nice for users and would have made a recent PR of @johannkm's nicer, who had to do some contortions in https://github.com/dagster-io/dagster/pull/16612 around `key_prefix`. 

This diff also adds documentation for `key` on `@asset`.

An upstack PR will consolidate the logic in `@asset` and `@graph_asset` around this stuff into a single helper function, since it is pretty gross, tricky code.

## How I Tested These Changes