Refactor asset key resolution in @asset and @graph_asset (#16752)
## Summary & Motivation

Takes the duplicated logic added in https://github.com/dagster-io/dagster/pull/16751 and consolidates it to a single helper function. There is a surprising amount of incidental complexity in this logic, so stuffing it into a common helper is a win.

I did this as a separate PR so that we can revert it independently in case this introduces a bug in `@asset`.

## How I Tested These Changes

BK