ext_protocol --> pipes_protocol (#16778)
## Summary & Motivation

This renames `ext_protocol` to `pipes_protocol`.

This also renames `ExtOrchestrationContext` to `PipesInvocation`. The reasoning here is that we are a bit "context happy" and we already have a lot of context objects. Instead, we just say this object represents an invocation you are about to do.

This argument against this name is that it closes off the possibility of multiple invocations in particular sessions.

Speaking of session, I did give that a go. However it felt too "heavy" for our use case now, which is doing a single invocation.

Update: respun this with `PipesSession` to try it out for size

## How I Tested These Changes

BK