[RFC] AssetLiveDataProvider (#16760)
## Summary & Motivation

Adding a new abstraction to solve the following problems:

Before:
- When an asset row scrolls out of view any in progress data fetching is
aborted even though the server is already doing the work
- When an asset row scrolls out of view we throw away the data that was
fetched, when it scrolls back into view we fetch it again (we do use the
apollo cache though!) even if the query is already cached and even if it
was last fetched recently.
- Asset fetches from different components are unbatched which creates
extra load on the server

After:
- Data from an asset fetch is stored in the apollo cache and the
requests are not aborted even if the component scrolls out of view
(Because aborting the request doesn't stop the server from doing the
work so we might aswell process the response and cache it in apollo)
- We track when assets are fetched globally which allows us to respect
the polling interval and not fetch an asset if it was recently fetched
- Asset fetches from different components are batched together, reducing
load on the server

Notably absent from this PR:
The logic that refetches everything when a Run is launched. I'll add
this in a follow-up. The end goal is for this hook to replace the
existing useLiveData hook.

## How I Tested These Changes

I mostly tested this locally by scrolling up/down in the OverviewAssets
root (http://localhost:4001/overview/activity/assets)

I will update the jest tests to be more comprehensive in a bit