ExtMessageReader --> PipesMessageReader
## Summary & Motivation

In the original RFC PR `ExtMessageReader` was renamed `PipeableMessageReader`. As in there are pipeable messages, and a reader is reading them. However others translated that class name to "there is a pipeable reader reading message", which ends up being very confusing.

For lack of any better ideas, I'm proposing just calling this `PipesMessageReader`, as in a message reader in the pipes library. I generally avoid class names as generic as `MessageReader`, as they are more likely to collide with classes elsewhere in the python environment and make things like auto-import features in IDEs less useful and more annoying. Unique names also help with greappability in the code base.

I'm open to more suggestions here. I thought about `InboundMessagePipe` or something, but I did not want to introduce pipe as a noun.

## How I Tested These Changes

BK