feat(asset-checks): allow asset check subsetting (#16672)
## Summary & Motivation
Build on https://github.com/dagster-io/dagster/pull/16610 and
https://github.com/dagster-io/dagster/pull/16219 to enable asset check
subselection. This change was painful and a doozy.

To wrangle this complexity in the future we need to:
- Encompass a succinct and interpretable version of `XXXSelection` to
seamlessly enable the asset selection and asset check selection use case
- Standardize our usage of `None`, `[]`, `{}`, when indicating that a
selection is defaulting to select all the objects (assets and asset
checks).
- Rather than using `None`, an explicit sentinel value should be used
(e.g. `XXXSelection.ALL`, so that the value indicates its usage.
Likewise, `[]` and `{}` should be replaced with `XXXSelection.EMPTY`).

## How I Tested These Changes
- When materializing a subsetted asset, there are four cases:
  - Materialize without a selection
  - Materialize with a selection of assets and checks
  - Materialize with only a selection of assets
  - Materialize with only a selection of checks
- Remove any usages of `AssetSelection` that break the new subsetting
invariants (e.g. https://github.com/dagster-io/dagster/pull/16638)
- Existing pytest
- Stack on dbt implementation

---------

Co-authored-by: Johann Miller <johann@elementl.com>