Support description in AssetSpec on multi_asset. Leave comment explaining muddled nature of this codepath (#16755)
## Summary & Motivation

This was super confusing as by inspection it appeared that `multi_asset`
did not support descriptions. However there are actually three sources
of truth for descriptions, one in `OutputDefinition`, one in
`NodeDefinition`, and one in the `descriptions_by_key`. This last is
best thought of as an "override."

This goes in the annals of "un-composability and ambiguously immutable
Ls". There should be a canonical source of truth of descriptions once we
are deep within the depths of `AssetsDefinition`. If the dictionary is
needed for perf reasons (which I doubt), we can build a memoized cache
of descriptions derived from the immutable data. When the immutable data
is passed to a new copy of `AssetsDefinition` when we need it, we can
recompute the memoized dict.

## How I Tested These Changes

BK