[dagster-pipes] fix issue with pipes + non-time-partitioned assets (#16860)
## Summary & Motivation

Previously, if you attempted to use pipes with any non-time-partitioned
asset (i.e. static or dynamic), you'd get an error when building the
context, as `context.partition_time_window` would raise an error.

This PR makes sure we only call that in cases where it will not raise an
error. A bit gross that we have to dig into the step execution context
to get the partitions_def, but it is what it is.

## How I Tested These Changes

The new unit test failed before this change, passes now.