[graphql] Resolver to terminate runs (#16707)
This PR adds a resolver to terminate a set of runs, intended to support
a "terminate all queued runs" button. Under the hood, it calls
`instance.optimize_for_webserver` to pool connections and reduce query
time. The implementation in this PR takes 3.5 min to terminate 5K runs,
compared to ~15 min without the `optimize_for_webserver` call.

The resolver accepts a list of run IDs. This is intended so that the
front end can call `instance.get_run_ids(cursor=..., limit=...)` to
fetch run IDs in batches, then pass this to the resolver. This enables
displaying a progress bar.