Unify multipartition partition mappings (#16561)
[Conversation in
slack](https://dagsterlabs-workspace.slack.com/archives/C03AC5P8Z1T/p1693926896649959)
raised feedback from Pete that we should allow single ->
multidimensional partition dependency relationships where the time
granularities differ. Other users have also raised similar feedback.

Existing state before this PR:
- `SingleToMultiPartitionMapping` enables single -> multidimensional
deps where the single-dimensional partitions def must be a dimension of
the multipartitions def
- `MultiPartitionMapping` contains logic to handle deps between
arbitrary multipartitions

This PR unifies the logic in these two partition mappings in a
`BaseMultiPartitionMapping` class. It basically replaces the logic in
`SingleToMultiPartitionMapping` to use the same logic as
`MultiPartitionMapping` to enable `SingleToMultiPartitionMapping` to
call other partition mappings. Then end state is that
`SingleToMultiPartitionMapping` and `MultiPartitionMapping` now function
the same, but just have different construction signatures for
ergonomics.

The only behavioral change in this PR is that we now auto-infer a
`TimeWindowPartitionMapping` relationship between a single-dimension
time partitions def and a multipartitions def with a time dimension,
when the time partitions defs are not the same.

In the future, the goal is to allow users to provide a partition mapping
to `SingleToMultiPartitionMapping` (i.e. if they wanted to have a
specific `TimeWindowPartitionMapping` dependency).